services/file-organizer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
aiofiles==23.2.0
python-magic==0.4.27
pathlib
pydantic-settings
zeep
PyMuPDF==1.22.5
